requirements.txt
click==8.1.3
cloudinary==1.41.0
Flask==2.3.2
gunicorn
itsdangerous==2.1.2
Jinja2==3.1.3
loguru==0.7.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
python-dotenv==0.20.0
Werkzeug==3.0.3
